requirements.txt
fastapi
uvicorn
openai
python-dotenv
requests
pytrends
pydantic